[TASK] Separate rst rendering related tests
