+++ tests/Functional/RstRendering/RootIndexFileTest.php
@@ -1,17 +1,16 @@
 <?php
 declare(strict_types=1);
 
-namespace NamelessCoder\FluidDocumentationGenerator\Tests\Functional;
+namespace NamelessCoder\FluidDocumentationGenerator\Tests\Functional\RstRendering;
 
 use NamelessCoder\FluidDocumentationGenerator\Data\DataFileResolver;
-use NamelessCoder\FluidDocumentationGenerator\Entity\Schema;
 use NamelessCoder\FluidDocumentationGenerator\Export\RstExporter;
 use NamelessCoder\FluidDocumentationGenerator\SchemaDocumentationGenerator;
 use org\bovigo\vfs\vfsStream;
 use org\bovigo\vfs\vfsStreamDirectory;
 use PHPUnit\Framework\TestCase;
 
-class Typo3BackendViewhelperGroupWithoutSubGroupsIndexRstDocumentationGenerationTest extends TestCase
+class RootIndexFileTest extends TestCase
 {
     /**
      * @var vfsStreamDirectory
@@ -22,36 +21,27 @@ class Typo3BackendViewhelperGroupWithoutSubGroupsIndexRstDocumentationGeneration
      * the generated file is compared against this fixture file
      * @var string
      */
-    private $fixtureFilePath = __DIR__ . '/../Fixtures/rendering/output/Documentation/typo3/backend/9.4/Link/Index.rst';
+    private $fixtureFilePath = __DIR__ . '/../../Fixtures/rendering/output/Documentation/Index.rst';
 
     /**
      * output of the generation process
      * @var string
      */
-    private $generatedFilePath = 'outputDir/public/typo3/backend/9.4/Link/Index.rst';
+    private $generatedFilePath = 'outputDir/public/Index.rst';
 
     protected function setUp()
     {
         $this->vfs = vfsStream::setup('outputDir');
         $this->vfs->addChild(vfsStream::newDirectory('cache'));
         $dataFileResolver = DataFileResolver::getInstance(vfsStream::url('outputDir'));
-        $dataFileResolver->setResourcesDirectory(__DIR__ . '/../../resources/');
-        $dataFileResolver->setSchemasDirectory(__DIR__ . '/../Fixtures/rendering/input/');
+        $dataFileResolver->setResourcesDirectory(__DIR__ . '/../../../resources/');
+        $dataFileResolver->setSchemasDirectory(__DIR__ . '/../../Fixtures/rendering/input/');
         $schemaDocumentationGenerator = new SchemaDocumentationGenerator(
             [
                 new RstExporter()
             ]
         );
         $schemaDocumentationGenerator->generateFilesForRoot();
-        foreach ($dataFileResolver->resolveInstalledVendors() as $vendor) {
-            $schemaDocumentationGenerator->generateFilesForVendor($vendor);
-            foreach ($vendor->getPackages() as $package) {
-                $schemaDocumentationGenerator->generateFilesForPackage($package);
-                foreach ($package->getVersions() as $version) {
-                    $schemaDocumentationGenerator->generateFilesForSchema(new Schema($version));
-                }
-            }
-        }
     }
 
     /**
@@ -68,7 +58,7 @@ class Typo3BackendViewhelperGroupWithoutSubGroupsIndexRstDocumentationGeneration
     public function includeClausePointsToSettingsCfg()
     {
         $output = file($this->vfs->getChild($this->generatedFilePath)->url());
-        $this->assertSame('.. include:: ../../../../Includes.txt' . PHP_EOL, $output[0]);
+        $this->assertSame('.. include:: Includes.txt' . PHP_EOL, $output[0]);
     }
 
     /**
@@ -79,7 +69,7 @@ class Typo3BackendViewhelperGroupWithoutSubGroupsIndexRstDocumentationGeneration
         $output = file($this->vfs->getChild($this->generatedFilePath)->url());
         // first line is include, then empty, then upper headline decoration, then text -> fourth line
         $index = 3;
-        $this->assertSame('link' . PHP_EOL, $output[$index]);
+        $this->assertSame('Fluid ViewHelper Documentation' . PHP_EOL, $output[$index]);
     }
 
     /**
@@ -100,22 +90,12 @@ class Typo3BackendViewhelperGroupWithoutSubGroupsIndexRstDocumentationGeneration
     /**
      * @test
      */
-    public function viewHelperCountIsIntegrated()
-    {
-        $output = file($this->vfs->getChild($this->generatedFilePath)->url());
-        $index = 7;
-        $this->assertSame('* 2 ViewHelpers documented' . PHP_EOL, $output[$index]);
-    }
-
-    /**
-     * @test
-     */
     public function tocTreeContainsSubDirectoriesAsExpected()
     {
         $output = file($this->vfs->getChild($this->generatedFilePath)->url());
-        $index = 13;
-        $this->assertSame('   EditRecord' . PHP_EOL, $output[$index]);
-        $this->assertSame('   NewRecord' . PHP_EOL, $output[$index + 1]);
+        $index = 11;
+        $this->assertSame('   typo3/backend/9.4/Index' . PHP_EOL, $output[$index]);
+        $this->assertSame('   typo3/backend/9.5/Index' . PHP_EOL, $output[$index + 1]);
     }
 
     /**
@@ -123,7 +103,7 @@ class Typo3BackendViewhelperGroupWithoutSubGroupsIndexRstDocumentationGeneration
      */
     public function generatedFileIsSameAsFixture()
     {
-        $this->assertSame(trim(file_get_contents($this->fixtureFilePath)),
-            trim(file_get_contents($this->vfs->getChild($this->generatedFilePath)->url())));
+        $this->assertSame(file_get_contents($this->fixtureFilePath),
+            file_get_contents($this->vfs->getChild($this->generatedFilePath)->url()));
     }
 }
